Check before cleaning up a LocalDataSet's file
This is needed because the LocalDataSet's file field can now be None
(to support uploading the file separately through the upload endpoint)

+++ apps/storybase_asset/models.py
@@ -587,9 +587,10 @@ def delete_dataset_file(sender, instance, **kwargs):
     This should be hooked up to the post_delete signal
 
     """
-    storage, path = instance.file.file.storage, instance.file.file.path
-    instance.file.delete();
-    storage.delete(path)
+    if instance.file:
+        storage, path = instance.file.file.storage, instance.file.file.path
+        instance.file.delete();
+        storage.delete(path)
 
 post_delete.connect(delete_dataset_file, sender=LocalDataSet)
